Force cast IDs to int in constructors

+++ lib/models/Bookmark.php
@@ -28,9 +28,9 @@ class Bookmark
      */
     private function __construct($id, $user_id, $listing_id, DateTime $added)
     {
-        $this->id = require_non_empty($id, "bookmark_id");
-        $this->user_id = require_non_empty($user_id, "user_id");
-        $this->listing_id = require_non_empty($listing_id, "listing_id");
+        $this->id = (int)require_non_empty($id, "bookmark_id");
+        $this->user_id = (int)require_non_empty($user_id, "user_id");
+        $this->listing_id = (int)require_non_empty($listing_id, "listing_id");
         $this->added = require_non_null($added, "added");
     }
